Refined Loop Spell Detector script
Made it so the line gets thicker when a loop is ready.  It still detects the wrong number of intersections with certain specific shapes.  I plan on re-writing the script to detect intersections as the line is drawn instead of after.  This will have the benefit of giving the player feedback as they're drawing.